+++ lesson_8_9/src/main/java/com/example/lesson_8/dto/ErrorResponse.java
@@ -1,0 +1,8 @@
+package com.example.lesson_8.dto;
+
+import lombok.Data;
+
+@Data
+public class ErrorResponse {
+    private final String message;
+}
