+++ lesson_8_9/src/main/resources/db/migration/V1_0_0__init.sql
